requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2==2.9.5
SQLAlchemy==1.4.43
Werkzeug==2.2.2

python-dotenv~=0.21.0